+++ app/models/site_page.rb
@@ -59,8 +59,8 @@ class SitePage < Page
         steps = {pages: %w[position title type map],
                  names: %w[Position  Title Type Map]}
       when ContentType::HOMEPAGE
-        steps = {pages: %w[open_content open_content_preview],
-                 names: ['Open Content', 'Open Content Preview']}
+        steps = {pages: %w[title open_content open_content_preview],
+                 names: ['Title', 'Open Content', 'Open Content Preview']}
     end
     steps
   end